requirements-test.txt
djangorestframework>=3.12
pytest
pytest-django